chore: move utils to utils
